KlaussCPU: switch C int from 64-bit to 32-bit
Match the standard C type model — int = 32, long/long long/void* = 64 —
so third-party code (Embench, test-suite kernels, file-format readers)
ports cleanly and ordinary int arithmetic exercises the LDIDX32/STIDX32,
MEMGET32/MEMSET32, and SEXTW/ZEXTW codegen paths as a first-class case.

* clang KlaussCPUTargetInfo: IntWidth=32; DataLayout n32:64.
* KlaussCPUTargetMachine: matching DataLayout + comment refresh.
* KlaussCPUCallingConv.td: prepend CCPromoteToType<i64> to CC and RetCC
  so sub-word args/returns are widened to GPR width before assignment;
  required because the GPR class is i64-only.
* runtime/libc.c: heap_get_top now returns long long; heap_init writes
  the 8-byte heap-start word at address 0 via uint64_t* (not int*).
* runtime/test_64bit.c: tests retyped per intent — T1–T7 use long long
  (they specifically exercise 64-bit shifts/multiplies); new T3b checks
  signed int wrap at INT32_MAX; new T10d sum6_mixed exercises i32→i64
  promotion at call boundaries; T14 split into 32-bit and 64-bit globals;
  T15 walks both an int[] (sizeof=4, dist=12) and a long long[] (sizeof=8,
  dist=24); T16 retyped to long long * since the libc heap is 8-byte slotted.
* CLAUDE.md / FPGA_FIXES.md: updated DataLayout strings and added a Step 28
  changelog entry.

+++ libc.c
@@ -4,7 +4,9 @@
 // All functions assume the KlaussCPU hardware memory model:
 //   - Physical memory is little-endian: byte at address X in bits[8*(X mod 4)+7:8*(X mod 4)].
 //   - LDIDX8/STIDX8 use LE-lane mapping, consistent with physical memory.
-//   - int = 64-bit (matching the rcc ABI).
+//   - int = 32-bit; long, long long, void* are all 64-bit.
+//     The hardware heap-header word at address 0 is 8 bytes wide, so anything
+//     reading or writing it must use a 64-bit-pointer cast.
 
 typedef unsigned long long uint64_t;
 typedef long long           int64_t;
@@ -94,7 +96,8 @@ void *memcpy(void *dst, void *src, int n) {
 //
 // The heap starts at _end (first byte after .bss) and grows upward.
 // heap_start is written to address 0x0000 (the hardware heap-header area)
-// so test programs can read it with  int *mem = (int*)0; hs = mem[0];
+// as a 64-bit value, so test programs read it with
+//   long long *mem = (long long*)0; hs = mem[0];
 
 extern char _end[];   // first byte past .bss, from linker script
 
@@ -106,9 +109,11 @@ static char *g_heap_top  = 0;
 static void heap_init(void) {
     g_heap_base = _end;
     g_heap_top  = _end;
-    // Write heap start into the hardware heap-header area at address 0.
-    int *addr0 = (int *)0;
-    addr0[0] = (int)(uint64_t)g_heap_base;
+    // Write the 64-bit heap-start value into the hardware heap-header area
+    // at address 0.  Must be a 64-bit store: the hardware reserves 8 bytes
+    // here, and `int` is now 32 bits.
+    uint64_t *addr0 = (uint64_t *)0;
+    addr0[0] = (uint64_t)g_heap_base;
 }
 
 static uint64_t align8(uint64_t n) {
@@ -178,9 +183,12 @@ void *realloc(void *ptr, uint64_t newsize) {
     return newptr;
 }
 
-// Return the current heap top as an integer address.
-int heap_get_top(void) {
-    return (int)(uint64_t)g_heap_top;
+// Return the current heap top as a 64-bit address.  Returning `int` would
+// truncate the (32-bit) address to the low 32 bits and sign-extend it to
+// 64 bits when read by callers — fine for valid 128 MiB addresses, but
+// confusing.  Use the unambiguous wider type.
+long long heap_get_top(void) {
+    return (long long)(uint64_t)g_heap_top;
 }
 
 // Return the total number of 8-byte words currently in-use by live allocations.
